app/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
django-rest-framework
django-debug-toolbar
django-haystack
Pillow==10.1.0
pysolr==3.9.0